Fix bug in allocation method and list

+++ kernel/alloc/liballoc.h
@@ -34,7 +34,7 @@ struct	boundary_tag
     unsigned int magic;			//< It's a kind of ...
     unsigned long long size; 			//< Requested size.
     unsigned long long real_size;		//< Actual size.
-    unsigned long long index;					//< Location in the page table.
+    int index;					//< Location in the page table.
 
     struct boundary_tag *split_left;	//< Linked-list info for broken pages.
     struct boundary_tag *split_right;	//< The same.
